Environment ready for project developments. The project code is not committed now because it is in an external directory.

+++ Source/Core/BareMetal/L1Portability/Environment/FreeRTOS/BasicConsole.cpp
@@ -147,7 +147,7 @@ uint64 BasicConsole::Size() {
 bool BasicConsole::Seek(uint64 pos) {
     return false;
 }
-bool BasicConsole::RelativeSeek(const int32 deltaPos) {
+bool BasicConsole::RelativeSeek(const int64 deltaPos) {
     return false;
 }
 uint64 BasicConsole::Position() {
