+++ Source/Core/BareMetal/L3Streams/BufferedStreamGenerator.h
@@ -51,6 +51,11 @@ namespace MARTe {
  *
  * All the other operating system stream functions will be left untouched allowing full access to the specific
  * operating system stream implementation.
+ *
+ * @tparam bufferedStream a class representing a buffered stream, i.e. a class
+ * derived from the abstract class BufferedStreamI.
+ * @tparam basicStream a class representing a basic stream, i.e. a class
+ * derived from the abstract class StreamI.
  */
 template<class bufferedStream, class basicStream>
 class BufferedStreamGenerator: public bufferedStream, public basicStream {
@@ -130,7 +135,7 @@ public:
      * @brief Wraps bufferedStream::RelativeSeek(*).
      * BufferedStreamI::RelativeSeek(*).
      */
-    virtual bool RelativeSeek(const int32 deltaPos);
+    virtual bool RelativeSeek(const int64 deltaPos);
 
     /**
      * @brief Wraps bufferedStream::Position(*).
@@ -177,7 +182,7 @@ protected:
     /**
      * @brief Wraps the basicStream::RelativeSeek(*)
      */
-    virtual bool OSRelativeSeek(int32 deltaPos);
+    virtual bool OSRelativeSeek(int64 deltaPos);
 
     /**
      * @brief Wraps the basicStream::Position(*)
@@ -263,7 +268,7 @@ bool BufferedStreamGenerator<bufferedStream, basicStream>::Seek(uint64 pos) {
 }
 
 template<class bufferedStream, class basicStream>
-bool BufferedStreamGenerator<bufferedStream, basicStream>::RelativeSeek(const int32 deltaPos) {
+bool BufferedStreamGenerator<bufferedStream, basicStream>::RelativeSeek(const int64 deltaPos) {
     return bufferedStream::RelativeSeek(deltaPos);
 }
 
@@ -303,7 +308,7 @@ bool BufferedStreamGenerator<bufferedStream, basicStream>::OSSeek(uint64 pos) {
 }
 
 template<class bufferedStream, class basicStream>
-bool BufferedStreamGenerator<bufferedStream, basicStream>::OSRelativeSeek(int32 deltaPos) {
+bool BufferedStreamGenerator<bufferedStream, basicStream>::OSRelativeSeek(int64 deltaPos) {
     return basicStream::RelativeSeek(deltaPos);
 }
